# bookbyte/bookbinder/docs/swagger_config.yaml
openapi: 3.0.0
info:
  title: Bookbyte API
  description: API for managing books, library books, and persons
  version: "1.0.0"
  contact:
    email: contact@bookbyte.com
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:8080'
    description: Local Development Server
paths:
  /books/:
    post:
      tags:
        - book
      summary: Create a new book
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Book created
        '400':
          description: Error in request
        '409':
          description: Book already exists
    get:
      tags:
        - book
      summary: Get list of all books
      operationId: getAllBooks
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '204':
          description: No books found
  /books/overwrite:
    post:
      summary: Overwrite all books
      operationId: overwriteBooks
      tags:
        - book
      requestBody:
        description: JSON array of books to overwrite the existing catalog
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: All books overwritten successfully
        '400':
          description: Invalid request data (e.g., invalid ISBN)
  /books/{isbn13}:
    parameters:
      - in: path
        name: isbn13
        required: true
        schema:
          type: string
    put:
      tags:
        - book

      summary: Create or update a book by its ISBN
      operationId: upsertBookByISBN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book updated
        '201':
          description: Book created
        '400':
          description: Invalid input
    get:
      tags:
        - book

      summary: Get book by ISBN
      operationId: getBookByIsbn13
      parameters:
        - in: path
          name: isbn13
          required: true
          schema:
            type: string
            description: Unique identifier of the library book
      responses:
        '200':
          description: Detailed information about the library book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Invalid request or error fetching book details
    patch:
      tags:
        - book

      summary: Update a book by its ISBN
      operationId: updateBookByISBN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book updated
        '400':
          description: Invalid input
        '404':
          description: Book not found
    delete:
      summary: Delete a book by ISBN
      operationId: deleteBookByISBN
      tags:
        - book

      responses:
        '200':
          description: Book deleted
        '400':
          description: Invalid input
        '404':
          description: Book not found
...
components:
  schemas:
    Book:
      type: object
      required:
        - isbn13
        - title
        - author
      properties:
        isbn13:
          type: string
          example: '978-3-16-148410-0'
        title:
          type: string
          example: 'Clean Code'
        author:
          type: string
          example: 'Robert C. Martin'
...